requirements.txt
torch~=2.0.1
numpy~=1.24.3
timm~=0.9.2
torchvision~=0.15.2
Pillow~=9.5.0
scipy~=1.9.1
scikit-learn~=1.2.2
PyYAML~=6.0
yacs~=0.1.8
termcolor~=2.3.0
tqdm~=4.65.0
opencv-python==4.5.5.62
matplotlib~=3.7.0
scikit-image~=0.20.0
datasets~=2.12.0
albumentations~=1.3.1